+++ R/mcp.R
@@ -2,8 +2,9 @@
 #'
 #' Given a list of linear segments, `mcp` infers the posterior
 #' distributions of the parameters of each segment as well as the change points
-#' between segments. All segments must regress on the same x-axis. Change points
-#' are forced to be ordered using truncation of the priors. You can run
+#' between segments. [See more details and worked examples on the mcp website](https://lindeloev.github.io/mcp/).
+#' All segments must regress on the same x-variable. Change
+#' points are forced to be ordered using truncation of the priors. You can run
 #' `fit = mcp(segments, sample=FALSE)` to avoid sampling and the need for
 #' data if you just want to get the priors (`fit$prior`), the JAGS code
 #' `fit$jags_code`, or the R function to simulate data (`fit$func_y`).
@@ -83,26 +84,35 @@
 #' # Define the segments that are separated by change points
 #' segments = list(
 #'   score ~ 1 + year,  # intercept + slope
-#'    1 ~ 0 + year,  # joined slope
-#'    1 ~ 0,  # joined plateau
-#'    1 ~ 1  # disjoined plateau
+#'    ~ 0 + year,  # joined slope
+#'    ~ 0,  # joined plateau
+#'    ~ 1  # disjoined plateau
 #' )
 #'
-#' # Start sampling
+#' # Sample and see results
 #' fit = mcp(segments, data)
+#' summary(fit)
 #'
 #' # Visual inspection of the results
 #' plot(fit)
 #' plot(fit, "combo")
 #'
+#' # Test a hypothesis
+#' hypothesis(fit, "cp_1 > 10")
+#'
 #' # Compare to a one-intercept-only model (no change points) with default prior
-#' segments2 = list(1 ~ 1)
+#' segments2 = list(score ~ 1)
 #' fit2 = mcp(segments2, data)  # fit another model here
 #' fit$loo = loo(fit)
 #' fit2$loo = loo(fit)
 #' loo_compare(fit, fit2)
 #'
-#' # Show all priors (not just those specified manually)
+#' # Sample the prior and inspect it using all the usual methods (prior predictive checks)
+#' fit_prior = mcp(segments, data, sample = "prior")
+#' summary(fit_prior)
+#' plot(fit_prior)
+#'
+#' # Show all priors. Default priors are added where you don't provide any
 #' fit$prior
 #'
 #' # Set priors and re-run
@@ -119,16 +129,9 @@
 #' )
 #' fit3 = mcp(segments, data, prior)
 #'
-#' # Do stuff with the parameter estimates
-#' fit$pars$model  # check out which parameters are inferred.
-#' library(tidybayes)
-#' spread_draws(fit$mcmc_post, cp_1, cp_2, int_1, year_1, year_2) %>%
-#'    # tidybayes stuff here
-#'
 #' # Show JAGS model
 #' cat(fit$jags_code)
 #' }
-
 
 mcp = function(segments,
                data = NULL,
@@ -217,6 +220,7 @@ mcp = function(segments,
 
   # Get an abstract segment table ("ST")
   ST = get_segment_table(segments, data, family$family, par_x)
+
   par_x = unique(ST$x)
   par_y = unique(ST$y)
   par_trials = unique(ST$trials)
@@ -226,11 +230,13 @@ mcp = function(segments,
   params_varying = logical0_to_null(c(stats::na.omit(ST$cp_group)))
   params_population = names(prior)[!names(prior) %in% params_varying]
   #params_population = c(stats::na.omit(unique(c(ST$int_name, ST$slope_name, ST$cp_name[-1], ST$cp_sd))))
-  # if (family$family == "gaussian")
-  #   params_population = c(params_population, "sigma")
 
   # Make formula_str and func_y
   formula_str = get_formula_str(ST, par_x)
+  if (family$family == "gaussian") {
+    formula_str_sigma = get_formula_str(ST, par_x, sigma = TRUE)
+    formula_str = paste0(formula_str, "\n\n", formula_str_sigma)
+  }
 
   params_funcy = params_population[!params_population %in% ST$cp_sd]
   func_y = get_func_y(formula_str, par_x, par_trials, params_funcy, params_varying, nrow(ST), family$family)
@@ -335,7 +341,6 @@ mcp = function(segments,
 #'@aliases logical0_to_null
 #'@param x Anything
 #'@return NULL or x
-
 
 logical0_to_null = function(x) {
   if (length(x) > 0)